Added test + small refac
